LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-fieldset-003-expected.txt





FAIL .target > * 1 assert_equals:
<div data-offset-x="25"><span></span></div>
offsetLeft expected 25 but got 0
FAIL .target > * 2 assert_equals:
<fieldset data-offset-x="0">
    <legend><span></span></legend>
    <span></span><br><span></span>

</fieldset>
offsetLeft expected 0 but got 5
FAIL .target > * 3 assert_equals:
<div data-offset-x="45"><span></span></div>
offsetLeft expected 45 but got 0
FAIL .target > * 4 assert_equals:
<fieldset data-offset-x="0">
    <span></span><br><span></span>
    <legend><span></span></legend>

</fieldset>
offsetLeft expected 0 but got 5
